Выносит список фильмов в компонент Каталог для уменьшения ререндеринга
